Timer system manipulates the visualization and notifications pretty much exactly how we want it to. No outstanding bugs, knock on wood.

+++ migrations/20160618140100_add_notified_at_to_cards.sql
@@ -19,7 +19,7 @@ CREATE TABLE cards (
   deck_id INTEGER NOT NULL,
   content_html TEXT NOT NULL,
   orbit INTEGER,
-  notified_at INTEGER,
+  notified_at BIGINT,
   created_at TIMESTAMP WITHOUT TIME ZONE NOT NULL,
   modified_at TIMESTAMP WITHOUT TIME ZONE NOT NULL
 );
